+++ sdk/identity/azure_identity/src/credentials/cache.rs
@@ -15,6 +15,7 @@ impl TokenCache {
         Self(RwLock::new(HashMap::new()))
     }
 
+    #[allow(dead_code)]
     pub(crate) async fn clear(&self) -> azure_core::Result<()> {
         let mut token_cache = self.0.write().await;
         token_cache.clear();